Extract funding domain: gas-station + wallet-activator; hedera probe to network
GasStation's mechanism is generic native-token funding — its only custody
ties were the CustodyWallet type and being provisioned by the provider.
Its consumers span plan-approval options, omnibus-delegate and the
deposit watchers, none of them custody-internal. Give funding its own
domain:

- services/funding/: gas-station (sender-side, recurring) and
  wallet-activator (recipient-side, one-time) — the documented sibling
  pair sharing the funding wallet and timeout constants. GasStation now
  takes a minimal FundingWallet {provider, signer}, which CustodyWallet
  satisfies structurally, so funding imports nothing from custody.
- services/network/: isHederaNetwork + HEDERA_CHAIN_IDS join health —
  network probes, previously bundled with the activator.
- services/custody/ is reduced to the actual custody boundary: provider
  interface + registry, with gasStation? typed from funding.
