+++ src/integrations/dfns/provider.ts
@@ -3,7 +3,8 @@ import { AsymmetricKeySigner } from '@dfns/sdk-keysigner';
 import { DfnsWallet } from '@dfns/lib-ethersjs6';
 import { JsonRpcProvider } from 'ethers';
 import { DfnsAppConfig } from './config';
-import { CustodyProvider, CustodyWallet, GasStation } from '../../services/custody';
+import { CustodyProvider, CustodyWallet } from '../../services/custody';
+import { GasStation } from '../../services/funding';
 
 export class DfnsCustodyProvider implements CustodyProvider {
   readonly escrow: CustodyWallet;
