+++ src/services/custody/index.ts
@@ -1,4 +1,2 @@
 export * from './custody-provider'
 export * from './custody-registry'
-export * from './gas-station'
-export * from './wallet-activation'
