+++ src/services/network/index.ts
@@ -1,1 +1,2 @@
 export * from './health'
+export * from './hedera'
