+++ src/services/plan-approval/options/wallet-activation-option.ts
@@ -2,7 +2,7 @@ import { logger } from "@owneraio/finp2p-nodejs-skeleton-adapter";
 import { PlanApprovalOption, IntrospectedPlan } from "..";
 import { AccountResolver } from "../../accounts/account-resolver";
 import { CustodyProvider } from "../../custody/custody-provider";
-import { DEFAULT_ACTIVATION_AMOUNT, WalletActivator } from "../../custody/wallet-activation";
+import { DEFAULT_ACTIVATION_AMOUNT, WalletActivator } from "../../funding/wallet-activation";
 
 /**
  * Plan-approval option that activates recipient wallets on Hedera-style
